refactor(console-ui): phase 5e — chat module + streaming perf (F3/F7/F10-any)
Split components/ChatMessage.tsx into a components/chat/ module (UserMessage,
ChatMessage thin orchestrator, MessageBody, ThinkingBlock, StreamMetrics,
CodeBlock, typed markdown.tsx) and fix the streaming re-render storm:

- React.memo(ChatMessage) + an id-keyed, referentially-stable onRetry so only
  the streaming message re-renders, not the whole list (perf F3).
- Render plain text while streaming and parse markdown once on completion —
  removes the O(n^2) re-tokenization of the growing reply (perf F7).
- New hooks/useChatStream.ts unifies the near-duplicate send + retry
  orchestration and batches per-token store writes with requestAnimationFrame
  (~one store update per frame instead of per token) (perf F3c). app/page.tsx is
  now a thin orchestrator.
- Narrow Zustand selectors in ChatInput / TopBar / Sidebar so they no longer
  re-render on every streamed token (perf F3).
- markdown.tsx is typed against react-markdown's Components (drops the 4 `any`,
  F10).

Invariant held: build green, eslint 0 errors / 113 warnings (down from 116),
vitest 352 pass / 4 pre-existing fail. Chat route 360 KB gz (runtime win; the
eager react-markdown/pkijs trim lands with verification phase F4).
